% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mlb_stadium.R
\name{geom_mlb_stadium}
\alias{geom_mlb_stadium}
\title{Geom MLB Stadium}
\usage{
geom_mlb_stadium(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = FALSE, stadium_ids = NULL,
  stadium_segments = "outfield_outer", stadium_transform_coords = FALSE,
  ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behavior from
the default plot specification, e.g. \code{\link{borders}}.}

\item{stadium_ids}{mapping to pass to `If \code{TRUE}, the default, will check that
supplied parameters and aesthetics are understood by the \code{geom} or
\code{stat}. Use \code{FALSE} to suppress the checks.}

\item{params}{Additional parameters to the \code{geom} and \code{stat}.}
}
\description{
This draws the outline of an MLB stadium.
}
\details{
The data are based on the SVG files used by MLBAM. The SVG files have been
converted to a \code{data.frame} for use with \code{ggplot2::geom_path}
}
\examples{
set.seed(101)
batted_ball_data = data.frame(hc_x = rnorm(20, 125, 10), hc_y=rnorm(20, 100, 20))
batted_ball_data$team = rep(c("angels", "yankees"), each=10)
g <- ggplot(batted_ball_data) + geom_spraychart(aes(x=hc_x, y=hc_y), stadium_ids = "angels")

}
\seealso{
\code{\link{geom_spraychart}}
}
