% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mlb_stadium.R
\name{geom_spraychart}
\alias{geom_spraychart}
\title{MLB Spraychart}
\usage{
geom_spraychart(mapping = NULL, data = NULL, stat = "identity",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, stadium_aes = NULL, stadium_ids = NULL,
  stadium_segments = "outfield_outer", stadium_transform_coords = FALSE,
  ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), it is combined with the default mapping at the top level of the
plot. You must supply \code{mapping} if there is no plot mapping.}

\item{data}{The data to be displayed in this layer. There are three
   options:

   If \code{NULL}, the default, the data is inherited from the plot
   data as specified in the call to \code{\link{ggplot}}.

   A \code{data.frame}, or other object, will override the plot
   data. All objects will be fortified to produce a data frame. See
   \code{\link{fortify}} for which variables will be created.

   A \code{function} will be called with a single argument,
   the plot data. The return value must be a \code{data.frame.}, and
   will be used as the layer data.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behavior from
the default plot specification, e.g. \code{\link{borders}}.}

\item{stadium_aes}{mapping to pass to \code{geom_mlb_stadium}}

\item{stadium_ids}{character Vector of stadium ids.}

\item{stadium_segments}{character Vector of \code{stadium_segments} to display. Options are
'outfield_inner', 'outfield_outer', 'infield', 'all'.}

\item{stadium_transform_coords}{boolean if \code{TRUE}, then transform the path data
according to \code{mlb_xy_transformation}}

\item{...}{other arguments passed on to \code{\link{layer}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{color = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
This generates a spraychart - a scatter plot of \code{x, y} values
overlayed on a baseball field. It effectively overrides
\code{geom_point} and additionally concatenates a \code{geom_mlb_stadium} layer.
}
\examples{
g = ggplot() + geom_mlb_
set.seed(101)
batted_ball_data = data.frame(hc_x = rnorm(20, 125, 10), hc_y=rnorm(20, 100, 20))
batted_ball_data$team = rep(c("angels", "yankees"), each=10)
g <- ggplot(batted_ball_data) + geom_spraychart()

}
\seealso{
\code{\link{geom_mlb_stadium}}
}
